requirements.txt
coinmarketcapapi
pycoingecko
pymongo
discord.py